requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scikit-image>=0.21
Pillow>=10.0
joblib>=1.3
# Optional, only for run_05_gpu_coates.py (GPU-accelerated patch encoding)
# cupy-cuda12x
